Fix update from 4.8.0-1 and earlier
In 4.8.0-2 the method to unbundle MathJax was changed from a simple
symlink handled by the %posttrans and %preun scriptlets to the one
implemented by Calibre itself. This made updating from 4.8.0-1 and
earlier impossible. Renaming the MathJax folder and letting the
scriptlets create a symlink again is ugly but fixes this issue.